encapsulando metodos de criacao de entidade e alterando tipo do waist

+++ src/main/resources/db/migration/structure/V1.4.0__create_table_body-measurement.sql
@@ -3,7 +3,7 @@ CREATE TABLE body_measurement (
     id                              varchar(36) NOT NULL,
     student_id                      varchar(36) NOT NULL,
     collection_date                 date NOT NULL,
-    waist                           decimal(5,2) NOT NULL,
+    waist                           integer NOT NULL,
     weight                          decimal(5,2) NOT NULL,
     height                          integer NOT NULL,
     bmi                             decimal(5,2) NOT NULL,
